requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.66.0
pandas>=1.5.0
google-generativeai>=0.2.0
xgboost>=1.7.0
joblib>=1.3.0
openpyxl>=3.1.2

